Remove Old DBConnect and sql file
The new DBConnect.java has the same functionallity as what
was in the DBConnect file, but it has several more methods
to retrieve and manipulate data from a database. Also it is
actually now a java file. where to place the new DBConnect
should be decided later when it is decided how to implement it.
For now, it will stay in the root of the project.

The new sql file ilearn.sql creates both the user and classes
tables and inserts some default items into them.